renamed locales.js file
